backend/pyproject.toml
[project]
name = "nemoguard-backend"
version = "0.1.0"
description = "NemoGuard backend modular monolith MVP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2,<2.0.0",
  "argon2-cffi>=25.1.0,<26.0.0",
  "celery[redis]>=5.4.0,<6.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "psycopg[binary]>=3.2.1,<4.0.0",
  "pydantic-settings>=2.5.2,<3.0.0",
  "pyjwt>=2.9.0,<3.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "python-multipart>=0.0.9,<0.1.0",
  "redis>=5.0.8,<6.0.0",
  "sqlalchemy>=2.0.35,<3.0.0",
  "uvicorn[standard]>=0.30.6,<1.0.0",
  "pwdlib>=0.2.0,<1.0.0",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2,<1.0.0",
  "pytest>=8.3.3,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
